Follow all suggested user accounts (IOS-157)
